chore: rename smoke test to match coding conventions
