+++ tests/smoke-tests/NHS.CohortManager.SmokeTests.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.002.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NHS.CohortManager.SmokeTests", "NHS.CohortManager.SmokeTests\NHS.CohortManager.SmokeTests.csproj", "{B0014134-D3E5-496A-941E-DE76EF985E4F}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{B0014134-D3E5-496A-941E-DE76EF985E4F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B0014134-D3E5-496A-941E-DE76EF985E4F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B0014134-D3E5-496A-941E-DE76EF985E4F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B0014134-D3E5-496A-941E-DE76EF985E4F}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A38D9FB5-421A-4CC8-86DB-40ECD0E25E64}
+	EndGlobalSection
+EndGlobal
