+++ tests/smoke-tests/NHS.CohortManager.SmokeTests/Config/FilePaths.cs
@@ -1,0 +1,8 @@
+namespace NHS.CohortManager.SmokeTests.Config;
+
+public class FilePaths
+{
+    public string Add { get; set; }
+    public string Amended { get; set; }
+    public string Remove { get; set; }
+}
