+++ tests/smoke-tests/NHS.CohortManager.SmokeTests/Config/appsettings.json
