+++ tests/smoke-tests/NHS.CohortManager.SmokeTests/Models/FieldsTable.cs
@@ -1,0 +1,7 @@
+namespace NHS.CohortManager.SmokeTests.Models;
+
+public class FieldsTable
+{
+    public string? FieldName { get; set; }
+    public string? FieldValue { get; set; }
+}
