+++ tests/smoke-tests/NHS.CohortManager.SmokeTests/Models/NHSNumbers.cs
@@ -1,0 +1,12 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace NHS.CohortManager.SmokeTests.Models;
+
+public class NhsNumbers
+{
+    public string? NhsNumber { get; set; }
+}
