+++ tests/smoke-tests/NHS.CohortManager.SmokeTests/Models/RecordTypesEnum.cs
@@ -1,0 +1,13 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace NHS.CohortManager.SmokeTests.Models;
+
+public enum RecordTypesEnum
+{
+    Add,
+    Amended
+}
