+++ tests/smoke-tests/NHS.CohortManager.SmokeTests/NHS.CohortManager.SmokeTests.csproj
@@ -1,0 +1,96 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net9.0</TargetFramework>
+        <RootNamespace>NHS.CohortManager.EndToEndTests</RootNamespace>
+        <LangVersion>latest</LangVersion>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <IsPackable>false</IsPackable>
+        <NoWarn>CS8618;CS8604;CS8605;CS8601;CS8604;CS8602;CS8625</NoWarn>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <Compile Remove="TestFiles\remove\**" />
+      <Compile Remove="TestFiles\update\**" />
+      <EmbeddedResource Remove="TestFiles\remove\**" />
+      <EmbeddedResource Remove="TestFiles\update\**" />
+      <None Remove="TestFiles\remove\**" />
+      <None Remove="TestFiles\update\**" />
+      <SpecFlowFeatureFiles Remove="TestFiles\remove\**" />
+      <SpecFlowFeatureFiles Remove="TestFiles\update\**" />
+      <SpecFlowObsoleteCodeBehindFiles Remove="TestFiles\remove\**" />
+      <SpecFlowObsoleteCodeBehindFiles Remove="TestFiles\update\**" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Azure.Identity" Version="1.13.2" />
+        <PackageReference Include="Azure.Storage.Blobs" Version="12.23.0" />
+        <PackageReference Include="ChoETL" Version="1.2.1.68" />
+        <PackageReference Include="ChoETL.Parquet" Version="1.0.1.34" />
+        <PackageReference Include="coverlet.collector" Version="6.0.2" />
+        <PackageReference Include="CsvHelper" Version="33.0.1" />
+        <PackageReference Include="FluentAssertions" Version="7.0.0" />
+        <PackageReference Include="Microsoft.Data.SqlClient" Version="5.2.2" />
+        <PackageReference Include="Microsoft.Extensions.Configuration" Version="9.0.0" />
+        <PackageReference Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="9.0.0" />
+        <PackageReference Include="Microsoft.Extensions.Configuration.FileExtensions" Version="9.0.0" />
+        <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="9.0.0" />
+        <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="9.0.0" />
+        <PackageReference Include="Microsoft.Extensions.Logging" Version="9.0.0" />
+        <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="9.0.0" />
+        <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="9.0.0" />
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.12.0" />
+        <PackageReference Include="NUnit" Version="4.3.0" />
+        <PackageReference Include="NUnit.Analyzers" Version="4.4.0">
+          <PrivateAssets>all</PrivateAssets>
+          <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+        </PackageReference>
+        <PackageReference Include="NUnit3TestAdapter" Version="4.6.0" />
+        <PackageReference Include="SolidToken.SpecFlow.DependencyInjection" Version="3.9.3" />
+        <PackageReference Include="SpecFlow" Version="3.9.74" />
+        <PackageReference Include="SpecFlow.NUnit" Version="3.9.74" />
+        <PackageReference Include="SpecFlow.Tools.MsBuild.Generation" Version="3.9.74" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <Using Include="NUnit.Framework" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <None Update="Config\appsettings.json">
+        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+      </None>
+      <None Update="Config\appsettings-local.json">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+      </None>
+      <None Update="TestFiles\add\ADD1_1B8F53_-_CAAS_BREAST_SCREENING_COHORT.parquet">
+        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+      </None>
+      <None Update="TestFiles\add\ADD_1B8F53_-_CAAS_BREAST_SCREENING_COHORT.parquet">
+        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+      </None>
+      <None Update="TestFiles\add\ADD_2_RECORDS_-_CAAS_BREAST_SCREENING_COHORT.parquet">
+        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+      </None>
+      <None Update="TestFiles\add\ADD_500_2_-_CAAS_BREAST_SCREENING_COHORT.parquet">
+        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+      </None>
+      <None Update="TestFiles\add\Exception_1B8F53_-_CAAS_BREAST_SCREENING_COHORT.parquet">
+        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+      </None>
+      <None Update="TestFiles\amended\AMENDED1_1B8F53_-_CAAS_BREAST_SCREENING_COHORT.parquet">
+        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+      </None>
+      <None Update="TestFiles\amended\AMENDED_1B8F53_-_CAAS_BREAST_SCREENING_COHORT.parquet">
+        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+      </None>
+    </ItemGroup>
+
+    <ItemGroup>
+      <SpecFlowFeatureFiles Update="Features\SmokeTests\FileUploadAndCohortDistribution.feature">
+        <Generator Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">SpecFlowSingleFileGenerator</Generator>
+        <LastGenOutput>FileUploadAndCohortDistribution.feature.cs</LastGenOutput>
+      </SpecFlowFeatureFiles>
+    </ItemGroup>
+</Project>
