--- tests/smoke-tests/dtos-cohort-manager-smoke-tests/Config/AppSettings.cs
@@ -1,20 +1,0 @@
-using dtos_cohort_manager_specflow.Config;
-public class AzureSettings
-{
-    public bool IsCloudEnvironment { get; set; }
-}
-
-public class AppSettings
-{
-    public ConnectionStrings ConnectionStrings { get; set; }
-    public AzureSettings AzureSettings { get; set; }
-    public string? ManagedIdentityClientId { get; set; }
-    public FilePaths FilePaths { get; set; }
-    public string BlobContainerName { get; set; }
-    public string AzureWebJobsStorage { get; set; }
-}
-
-public class ConnectionStrings
-{
-    public string DtOsDatabaseConnectionString { get; set; }
-}
