--- tests/smoke-tests/dtos-cohort-manager-smoke-tests/Config/FilePaths.cs
@@ -1,8 +1,0 @@
-namespace dtos_cohort_manager_specflow.Config;
-
-public class FilePaths
-{
-    public string Add { get; set; }
-    public string Amended { get; set; }
-    public string Remove { get; set; }
-}
