--- tests/smoke-tests/dtos-cohort-manager-smoke-tests/Models/FieldsTable.cs
@@ -1,7 +1,0 @@
-namespace dtos_cohort_manager_specflow.Models;
-
-public class FieldsTable
-{
-    public string? FieldName { get; set; }
-    public string? FieldValue { get; set; }
-}
