--- tests/smoke-tests/dtos-cohort-manager-smoke-tests/Models/NHSNumbers.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace dtos_cohort_manager_specflow.Models;
-
-public class NhsNumbers
-{
-    public string? NhsNumber { get; set; }
-}
