--- tests/smoke-tests/dtos-cohort-manager-smoke-tests/Models/RecordTypesEnum.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace dtos_cohort_manager_specflow.Models;
-
-public enum RecordTypesEnum
-{
-    Add,
-    Amended
-}
