--- tests/smoke-tests/dtos-cohort-manager-smoke-tests/Models/TestRecords.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace dtos_cohort_manager_specflow.Models;
-
-public class TestRecords
-{
-    public string FileName { get; set; }
-    public RecordTypesEnum RecordType { get; set; }
-}
